pyproject.toml
[tool.poetry]
name = "cutethulhu"
version = "0.1.0"
description = "A bot for playing Call of Cthulhu via Discord"
authors = ["Dale Myers <dale@myers.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"
d20 = "^1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
